Splitting protocols and issue with documentation
