+++ Sources/QminderAPIProtocol.swift
@@ -24,25 +24,12 @@ public protocol QminderAPIProtocol {
   /// - Parameter completion: Callback block what is executed when location list is received
   func getLocationsList(completion: @escaping (Result<[Location], QminderError>) -> Void)
   
-  /// Get location list with response
-  ///
-  /// - Parameter completion: Callback block what is executed when location list is received
-  func getLocationsList(completion: @escaping (Result<[Location], QminderError>, HTTPURLResponse?) -> Void)
-  
   /// Get location details
   ///
   /// - Parameters:
   ///   - locationId: Location ID
   ///   - completion: Callback block what is executed when location data is received
   func getLocationDetails(locationId: Int, completion: @escaping (Result<Location, QminderError>) -> Void)
-  
-  /// Get location details with response
-  ///
-  /// - Parameters:
-  ///   - locationId: Location ID
-  ///   - completion: Callback block what is executed when location data is received
-  func getLocationDetails(locationId: Int,
-                          completion: @escaping (Result<Location, QminderError>, HTTPURLResponse?) -> Void)
   
   /// Get Location Lines
   ///
@@ -51,28 +38,12 @@ public protocol QminderAPIProtocol {
   ///   - completion: Callback block what is executed when location lines is received
   func getLocationLines(locationId: Int, completion: @escaping (Result<[Line], QminderError>) -> Void)
   
-  /// Get Location Lines with response
-  ///
-  /// - Parameters:
-  ///   - locationId: Location ID
-  ///   - completion: Callback block what is executed when location lines is received
-  func getLocationLines(locationId: Int,
-                        completion: @escaping (Result<[Line], QminderError>, HTTPURLResponse?) -> Void)
-  
   /// Get location users
   ///
   /// - Parameters:
   ///   - locationId: Location ID
   ///   - completion: Callback block what is executed when location users is received
   func getLocationUsers(locationId: Int, completion: @escaping (Result<[User], QminderError>) -> Void)
-  
-  /// Get location users with response
-  ///
-  /// - Parameters:
-  ///   - locationId: Location ID
-  ///   - completion: Callback block what is executed when location users is received
-  func getLocationUsers(locationId: Int,
-                        completion: @escaping (Result<[User], QminderError>, HTTPURLResponse?) -> Void)
   
   /// Get location desks
   ///
@@ -81,14 +52,6 @@ public protocol QminderAPIProtocol {
   ///   - completion: Callback block what is executed when location users is received
   func getLocationDesks(locationId: Int, completion: @escaping (Result<[Desk], QminderError>) -> Void)
   
-  /// Get location desks with response
-  ///
-  /// - Parameters:
-  ///   - locationId: Location ID
-  ///   - completion: Callback block what is executed when location users is received
-  func getLocationDesks(locationId: Int,
-                        completion: @escaping (Result<[Desk], QminderError>, HTTPURLResponse?) -> Void)
-  
   // MARK: - Lines
   /// Get line details
   ///
@@ -96,14 +59,6 @@ public protocol QminderAPIProtocol {
   ///   - lineId: Line ID
   ///   - completion: Callback block what is executed when line data is received
   func getLineDetails(lineId: Int, completion: @escaping (Result<Line, QminderError>) -> Void)
-  
-  /// Get line details with response
-  ///
-  /// - Parameters:
-  ///   - lineId: Line ID
-  ///   - completion: Callback block what is executed when line data is received
-  func getLineDetails(lineId: Int,
-                      completion: @escaping (Result<Line, QminderError>, HTTPURLResponse?) -> Void)
   
   // MARK: - Tickets
   // swiftlint:disable function_parameter_count
@@ -128,42 +83,12 @@ public protocol QminderAPIProtocol {
                      limit: Int?, order: String?, responseScope: Set<String>?,
                      completion: @escaping (Result<[Ticket], QminderError>) -> Void)
   
-  // swiftlint:disable function_parameter_count
-  /// Search tickets with response
-  ///
-  /// - Parameters:
-  ///   - locationId: Optional parameter for searching tickets in specified location
-  ///   - lineId: Optional array of line ID's
-  ///   - status: Optional array of enum statuses
-  ///   - callerId: Optional parameter for searching tickets which were called by specified user ID
-  ///   - minCreatedTimestamp: Optional parameter for searching tickets which are created after specified time
-  ///   - maxCreatedTimestamp: Optional parameter for searching tickets which are created before specified time
-  ///   - minCalledTimestamp: Optional parameter for searching tickets which are called after specified time
-  ///   - maxCalledTimestamp: Optional parameter for searching tickets which are called before specified time
-  ///   - limit: Optional parameter for limiting number of search results. If no limit is specified, 1000 will be used
-  ///   - order: Optional parameter for ordering results
-  ///   - responseScope: Optional parameter for additional details about the found tickets
-  ///   - completion: Callback block executed when tickets are received back
-  func searchTickets(locationId: Int?, lineId: Set<Int>?, status: Set<Status>?,
-                     callerId: Int?, minCreatedTimestamp: Int?, maxCreatedTimestamp: Int?,
-                     minCalledTimestamp: Int?, maxCalledTimestamp: Int?,
-                     limit: Int?, order: String?, responseScope: Set<String>?,
-                     completion: @escaping (Result<[Ticket], QminderError>, HTTPURLResponse?) -> Void)
-  
   /// Get ticket details
   ///
   /// - Parameters:
   ///   - ticketId: Ticket ID
   ///   - completion: Callback block when ticket details are received
   func getTicketDetails(ticketId: String, completion: @escaping (Result<Ticket, QminderError>) -> Void)
-  
-  /// Get ticket details with response
-  ///
-  /// - Parameters:
-  ///   - ticketId: Ticket ID
-  ///   - completion: Callback block when ticket details are received
-  func getTicketDetails(ticketId: String,
-                        completion: @escaping (Result<Ticket, QminderError>, HTTPURLResponse?) -> Void)
   
   // MARK: - Users
   /// Get user details
@@ -173,23 +98,11 @@ public protocol QminderAPIProtocol {
   ///   - completion: Callback block when user details are received
   func getUserDetails(userId: Int, completion: @escaping (Result<User, QminderError>) -> Void)
   
-  /// Get user details with response
-  ///
-  /// - Parameters:
-  ///   - userId: User ID
-  ///   - completion: Callback block when user details are received
-  func getUserDetails(userId: Int, completion: @escaping (Result<User, QminderError>, HTTPURLResponse?) -> Void)
-  
   // MARK: - Devices
   /// Get pairing code and secred key
   ///
   /// - Parameter completion: Callback block what is executed when pairing code and secret key is received from server
   func getPairingCodeAndSecret(completion: @escaping (Result<TVPairingCode, QminderError>) -> Void)
-  
-  /// Get pairing code and secred key with response
-  ///
-  /// - Parameter completion: Callback block what is executed when pairing code and secret key is received from server
-  func getPairingCodeAndSecret(completion: @escaping (Result<TVPairingCode, QminderError>, HTTPURLResponse?) -> Void)
   
   /// Pair TV with code
   ///
@@ -199,28 +112,12 @@ public protocol QminderAPIProtocol {
   ///   - completion: Callback block when pairing is done on server:
   func pairTV(code: String, secret: String, completion: @escaping (Result<TVAPIData, QminderError>) -> Void)
   
-  /// Pair TV with code with return response
-  ///
-  /// - Parameters:
-  ///   - code: Pairing code
-  ///   - secret: Secret key
-  ///   - completion: Callback block when pairing is done on server:
-  func pairTV(code: String, secret: String,
-              completion: @escaping (Result<TVAPIData, QminderError>, HTTPURLResponse?) -> Void)
-  
   /// Get a details of a TV
   ///
   /// - Parameters:
   ///   - id: TV ID
   ///   - completion: Callback block when TV details are received
   func tvDetails(id: Int, completion: @escaping (Result<TVDevice, QminderError>) -> Void)
-  
-  /// Get a details of a TV with response
-  ///
-  /// - Parameters:
-  ///   - id: TV ID
-  ///   - completion: Callback block when TV details are received
-  func tvDetails(id: Int, completion: @escaping (Result<TVDevice, QminderError>, HTTPURLResponse?) -> Void)
   
   /// Update the heartbeat of the TV and add optional metadata in JSON format
   ///
@@ -232,16 +129,6 @@ public protocol QminderAPIProtocol {
                    metadata: [String: Any],
                    completion: @escaping (Result<Void?, QminderError>) -> Void)
   
-  /// Update the heartbeat of the TV and add optional metadata in JSON format with response
-  ///
-  /// - Parameters:
-  ///   - id: TV ID
-  ///   - metadata: Dictionary of metadata to send with heartbeat
-  ///   - completion: Callback block when TV heartbeat is received
-  func tvHeartbeat(id: Int,
-                   metadata: [String: Any],
-                   completion: @escaping (Result<Void?, QminderError>, HTTPURLResponse?) -> Void)
-  
   /// Empty state text message for TV
   ///
   /// - Parameters:
@@ -251,16 +138,6 @@ public protocol QminderAPIProtocol {
   func tvEmptyState(id: Int,
                     language: String,
                     completion: @escaping (Result<EmptyState, QminderError>) -> Void)
-  
-  /// Empty state text message for TV with response
-  ///
-  /// - Parameters:
-  ///   - id: TV ID
-  ///   - language: Language
-  ///   - completion: Callback block when empty state is received
-  func tvEmptyState(id: Int,
-                    language: String,
-                    completion: @escaping (Result<EmptyState, QminderError>, HTTPURLResponse?) -> Void)
 }
 
 extension QminderAPIProtocol {
@@ -272,24 +149,17 @@ extension QminderAPIProtocol {
                             limit: Int? = nil, order: String? = nil,
                             responseScope: Set<String>? = nil,
                             completion: @escaping (Result<[Ticket], QminderError>) -> Void) {
-    searchTickets(locationId: locationId, lineId: lineId,
-                  status: status, callerId: callerId,
-                  minCreatedTimestamp: minCreatedTimestamp, maxCreatedTimestamp: maxCreatedTimestamp,
-                  minCalledTimestamp: minCalledTimestamp, maxCalledTimestamp: maxCalledTimestamp,
-                  limit: limit, order: order, responseScope: responseScope, completion: completion)
-  }
-  
-  public func searchTickets(locationId: Int? = nil, lineId: Set<Int>? = nil,
-                            status: Set<Status>? = nil, callerId: Int? = nil,
-                            minCreatedTimestamp: Int? = nil, maxCreatedTimestamp: Int? = nil,
-                            minCalledTimestamp: Int? = nil, maxCalledTimestamp: Int? = nil,
-                            limit: Int? = nil, order: String? = nil,
-                            responseScope: Set<String>? = nil,
-                            completion: @escaping (Result<[Ticket], QminderError>, HTTPURLResponse?) -> Void) {
-    searchTickets(locationId: locationId, lineId: lineId,
-                  status: status, callerId: callerId,
-                  minCreatedTimestamp: minCreatedTimestamp, maxCreatedTimestamp: maxCreatedTimestamp,
-                  minCalledTimestamp: minCalledTimestamp, maxCalledTimestamp: maxCalledTimestamp,
-                  limit: limit, order: order, responseScope: responseScope, completion: completion)
+    searchTickets(locationId: locationId,
+                  lineId: lineId,
+                  status: status,
+                  callerId: callerId,
+                  minCreatedTimestamp: minCreatedTimestamp,
+                  maxCreatedTimestamp: maxCreatedTimestamp,
+                  minCalledTimestamp: minCalledTimestamp,
+                  maxCalledTimestamp: maxCalledTimestamp,
+                  limit: limit,
+                  order: order,
+                  responseScope: responseScope,
+                  completion: completion)
   }
 }
